+++ src/components/common/UserInvite.tsx
@@ -1,18 +1,18 @@
 import { FC } from 'react';
 import DialogButton from '@/components/common/DialogButton';
 import IconUserPlus from '@/assets/icons/IconUserPlus';
-import UserPlusDialog from '@/components/common/UserPlusDialog';
+import UserInviteDialog from '@/components/common/UserInviteDialog';
 
 const AddCalendarPage: FC = () => {
   return (
     <div>
       멤버 초대하기 *
       <DialogButton
-        classname={'userplus'}
+        classname={'userInvite bg-base-200 hover:bg-base-300'}
         name={<IconUserPlus />}
         title={'멤버 찾기'}
         desc={''}
-        children={<UserPlusDialog />}
+        children={<UserInviteDialog />}
       />
     </div>
   );
