+++ src/styles/index.css
@@ -2,12 +2,12 @@
 @tailwind components;
 @tailwind utilities;
 
-.userplus {
-  padding: 2rem 1.3rem;
+.userInvite {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 4.2rem;
+  height: 5.7rem;
   border: 1px solid oklch(var(--b3));
   border-radius: 0.5rem;
-  background-color: oklch(var(--b2));
-}
-.userplus:hover {
-  background-color: oklch(var(--b3));
 }
